# fix(centos7): add correct pymysql package on cent7/saltpy3