--- a/.gitlab-ci.yml
+++ b/.gitlab-ci.yml
@@ -138,8 +138,8 @@ default-debian-9-master-py3: {extends: '.test_instance'}
 default-ubuntu-2004-master-py3: {extends: '.test_instance'}
 default-ubuntu-1804-master-py3: {extends: '.test_instance'}
 default-ubuntu-1604-master-py3: {extends: '.test_instance'}
-# default-centos-8-master-py3: {extends: '.test_instance'}
-# default-centos-7-master-py3: {extends: '.test_instance'}
+default-centos-8-master-py3: {extends: '.test_instance'}
+default-centos-7-master-py3: {extends: '.test_instance'}
 # default-fedora-34-master-py3: {extends: '.test_instance'}
 # default-fedora-33-master-py3: {extends: '.test_instance'}
 # default-fedora-32-master-py3: {extends: '.test_instance'}