--- a/.yamllint
+++ b/.yamllint
@@ -20,6 +20,7 @@ ignore: |
   .kitchen/
   kitchen.vagrant.yml
   mysql/supported_sections.yaml
+  mysql/osfingermap.yaml
 
 yaml-files:
   # Default settings