--- a/test/integration/default/controls/packages_spec.rb
+++ b/test/integration/default/controls/packages_spec.rb
@@ -6,6 +6,8 @@ if (os[:name] == 'suse') || (os[:name] == 'opensuse')
   package_name = 'mariadb'
 elsif (os[:name] == 'debian') && os[:release].start_with?('8')
   package_name = 'mysql-server'
+elsif (os[:name] == 'centos') && os[:release].start_with?('8')
+  package_name = 'mysql-server'
 end
 
 control 'mysql package' do