### PR progress checklist (to be filled in by reviewers)
<!-- Please leave this checklist for reviewers to tick as they work through the PR. -->

- [x] Changes to documentation are appropriate (or tick if not required)
- [x] Changes to tests are appropriate (or tick if not required)
- [ ] Reviews completed

---

### What type of PR is this?
<!-- Please tick each box that is relevant (after creating the PR). -->

#### Primary type
<!-- There really should be only *one* of these types ticked for each PR. -->

- [ ] `[build]`    Changes related to the build system
- [ ] `[chore]`    Changes to the build process or auxiliary tools and libraries such as documentation generation
- [ ] `[ci]`       Changes to the continuous integration configuration
- [ ] `[feat]`     A new feature
- [x] `[fix]`      A bug fix
- [ ] `[perf]`     A code change that improves performance
- [ ] `[refactor]` A code change that neither fixes a bug nor adds a feature
- [ ] `[revert]`   A change used to revert a previous commit
- [ ] `[style]`    Changes that do not affect the meaning of the code (white-space, formatting, missing semi-colons, etc.)

#### Secondary type
<!-- Most PRs should include all of the following types as well. -->

- [ ] `[docs]`     Documentation changes
- [ ] `[test]`     Adding missing or correcting existing tests

### Does this PR introduce a `BREAKING CHANGE`?
<!-- If so, change the following to a `Yes` and explain what the breaking changes are. -->
<!-- If there are multiple breaking changes, list them all. -->

No.

### Related issues and/or pull requests
<!-- Please link any related issues/PRs here, especially any issues that are closed by this PR. -->

https://github.com/saltstack/salt/issues/54702
https://github.com/saltstack-formulas/mysql-formula/pull/248
https://github.com/saltstack-formulas/mysql-formula/pull/252

### Describe the changes you're proposing
<!-- A clear and concise description of what you have implemented. -->
<!-- Consider explaining each commit if they cover different aspects of the proposed changes. -->

**Problem:**
On Centos7 the `MySQL-python` package provides python2.7 support for MySQL. 
However, if salt is installed using python3 (salt-bootstrap -x python3) the salt MySQL state requires python3.6 support.
The result is that this formula fails on CentOS7.

**A solution**
This PR checks python version grain to choose the correct package to install on CentOS7.  
The solution is not ideal but I see no other solution for CentOS7.

### Pillar / config required to test the proposed changes
<!-- Provide links to the SLS files and/or relevant configs (be sure to remove sensitive info). -->



### Debug log showing how the proposed changes work
<!-- Include a debug log showing how these changes work, e.g. using `salt-minion -l debug`. -->
<!-- Alternatively, linking to Kitchen debug logs is useful, e.g. via. Travis CI. -->
<!-- Most useful is providing a passing InSpec test, which can be used to verify any proposed changes. -->

*** BEFORE ***

```
          ID: mysql_python
    Function: pkg.installed
        Name: MySQL-python
      Result: True
     Comment: The following packages were installed/updated: MySQL-python
     Started: 21:35:15.551836
    Duration: 3358.682 ms
     Changes:
              ----------
              MySQL-python:
                  ----------
                  new:
                      1.2.5-1.el7
                  old:
----------
          ID: mysql_delete_anonymous_user_localhost
    Function: mysql_user.absent
        Name:
      Result: False
     Comment: State 'mysql_user.absent' was not found in SLS 'mysql.server'
              Reason: 'mysql_user' __virtual__ returned False: mysql module could not be loaded
     Changes:
----------
          ID: mysql_delete_anonymous_user_localhost.localdomain
    Function: mysql_user.absent
        Name:
      Result: False
     Comment: State 'mysql_user.absent' was not found in SLS 'mysql.server'
              Reason: 'mysql_user' __virtual__ returned False: mysql module could not be loaded
     Changes:
```

** VERSIONS**
```
[root@localhost saltstack-formulas]# salt-call grains.items --local | grep -A2 pythonversion
    pythonversion:
        - 3
        - 6
        
[root@localhost saltstack-formulas]# cat /etc/redhat-release
CentOS Linux release 7.9.2009 (Core)

[root@localhost saltstack-formulas]# salt --version
salt 3003
```

** Try python3-pyMySQL  https://github.com/saltstack-formulas/mysql-formula/pull/252   **
```
        ID: mysql_python
    Function: pkg.installed
        Name: python3-PyMySQL
      Result: False
     Comment: The following packages failed to install/update: python3-PyMySQL
     Started: 22:46:20.552165
    Duration: 14220.153 ms
     Changes:
              ----------
              python36-PyMySQL:
                  ----------
                  new:
                      0.9.3-1.el7
                  old:
----------
          ID: mysql_delete_anonymous_user_localhost
    Function: mysql_user.absent
        Name:
      Result: False
     Comment: One or more requisite failed: mysql.python.mysql_python
     Started: 22:46:36.629242
    Duration: 0.003 ms
     Changes:
----------
          ID: mysql_delete_anonymous_user_localhost.localdomain
    Function: mysql_user.absent
        Name:
      Result: False
     Comment: One or more requisite failed: mysql.python.mysql_python
     Started: 22:46:36.629510
    Duration: 0.002 ms
```

**  Try python36-mysql **
```
          ID: mysql_python
    Function: pkg.installed
        Name: python36-mysql
      Result: True
     Comment: The following packages were installed/updated: python36-mysql
     Started: 22:30:09.846993
    Duration: 15459.45 ms
     Changes:
              ----------
              python36-mysql:
                  ----------
                  new:
                      1.3.12-2.el7
                  old:
----------
          ID: mysql_delete_anonymous_user_localhost
    Function: mysql_user.absent
        Name:
      Result: True
     Comment: User @localhost is not present, so it cannot be removed
     Started: 22:30:26.883642
    Duration: 8.302 ms
     Changes:
----------
          ID: mysql_delete_anonymous_user_localhost.localdomain
    Function: mysql_user.absent
        Name:
      Result: True
     Comment: User @localhost.localdomain is not present, so it cannot be removed
     Started: 22:30:26.892263
    Duration: 3.298 ms
```
** Try python36-PyMySQL **
```
          ID: mysql_python
    Function: pkg.installed
        Name: python36-PyMySQL
      Result: True
     Comment: The following packages were installed/updated: python36-PyMySQL
     Started: 22:52:41.540460
    Duration: 21308.042 ms
     Changes:
              ----------
              python36-PyMySQL:
                  ----------
                  new:
                      0.9.3-1.el7
                  old:
---------
          ID: mysql_delete_anonymous_user_localhost
    Function: mysql_user.absent
        Name:
      Result: True
     Comment: User @localhost is not present, so it cannot be removed
     Started: 22:53:04.420969
    Duration: 11.397 ms
     Changes:
----------
          ID: mysql_delete_anonymous_user_localhost.localdomain
    Function: mysql_user.absent
        Name:
      Result: True
     Comment: User @localhost.localdomain is not present, so it cannot be removed
     Started: 22:53:04.432724
    Duration: 4.35 ms

---------

### Documentation checklist
<!-- Please tick each box that is relevant (after creating the PR). -->

- [ ] Updated the `README` (e.g. `Available states`).
- [ ] Updated `pillar.example`.

### Testing checklist
<!-- Please tick each box that is relevant (after creating the PR). -->

- [ ] Included in Kitchen (i.e. under `state_top`).
- [ ] Covered by new/existing tests (e.g. InSpec, Serverspec, etc.).
- [ ] Updated the relevant test pillar.

### Additional context
<!-- Add any other context about the proposed changes here. -->