@noelmcloughlin 2021-06-10
I managed to get centos CI working.