@ixs 2021-07-12
Just a note, this might be a breaking change for existing systems. If anyone is still running salt on the system default python 2.7, it would break.

Maybe wrap the logic in a something like this:

```
{%- if salt["grains.get"]("pythonversion")[0] == 2 %}
  foo
{%- else %}
  bar
{%- endif %}
```