@noelmcloughlin 2021-08-09
> Maybe wrap the logic in a something like this:

Done.

```
     {%- if 'pythonversion' in grains and grains.pythonversion[0]|int == 3 %}
  pythonpkg: python36-PyMySQL   # python36-mysql works too
      {%- else %}
  pythonpkg: MySQL-python
      {%- endif %}
```