@saltstack-formulas-travis 2021-09-04
:tada: This PR is included in version 0.56.2 :tada:

The release is available on [GitHub release](https://github.com/saltstack-formulas/mysql-formula/releases/tag/v0.56.2)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: